Refactored the base classes to allow unit testing of Jenkins jobs.
